moving association functions to separate file
